requirements.txt
opencv-python
Kivy
numpy
pillow
dlib
scipy
pyqt5